Typing and status functions implemented.

+++ toxcore/Messenger.c
@@ -806,7 +806,7 @@ int m_set_usertyping(Messenger *m, int32_t friendnumber, uint8_t is_typing)
     return 0;
 }
 
-uint8_t m_get_istyping(const Messenger *m, int32_t friendnumber)
+int m_get_istyping(const Messenger *m, int32_t friendnumber)
 {
     if (friend_not_valid(m, friendnumber))
         return -1;
